refactor(#67): Move list component item to other file
- move it to a separated file
